[014] Add Season Styling on League Show
This adds styling on league#show for the active_season.
It also extracts the season table row on league#show page to its own
partial. This page is going to get big from all of the data being
rendered on it and staying proactive with partials would be a great
approach.