Add spec to remove player from game, add ordering for finishing place.
I think a better approach for future features like moving player up and
down the list of finishers is better by just taking the order they are
in. I was originally going to swap timestamps, but comparing timestamps
seems more complicated than adding a number to a player and just sorting
that way.